created a separate file for RTT functions (#39)
* created a separate file for RTT functions

* added new rtt file to base s2s file
