+++ brainclouds2s.js
@@ -1,6 +1,8 @@
 var https = require('https')
 var util = require('util')
 
+RTT = require('./brainclouds2s-rtt')
+
 // Constants
 const SERVER_SESSION_EXPIRED = 40365    // Error code for expired session
 const HEARTBEAT_INTERVALE_MS = 60 * 30 * 1000   // 30 minutes heartbeat interval
@@ -8,27 +10,6 @@ const HEARTBEAT_INTERVALE_MS = 60 * 30 * 1000   // 30 minutes heartbeat interval
 const STATE_DISCONNECTED = 0
 const STATE_AUTHENTICATING = 1
 const STATE_CONNECTED = 2
-
-// RTT
-var socket = null
-
-var rttConnectionState = {
-    CONNECTED: "Connected",
-    DISCONNECTED: "Disconnected",
-    CONNECTING: "Connecting",
-    DISCONNECTING: "Disconnecting"
-}
-var rttConnectionStatus = rttConnectionState.DISCONNECTED
-var rttConnectCallback = null
-var DEFAULT_RTT_HEARTBEAT // Seconds
-var rttHeartbeatId = null
-var disconnectedWithReason = false
-var disconnectMessage = null
-var rttAuth = {}
-var rttCallback = null
-var debugEnabled = true
-var appId = null
-var sessionId = null
 
 // Execute the S2S request
 function s2sRequest(context, json, callback) {
@@ -99,26 +80,6 @@ function startHeartbeat(context) {
             }
         })
     }, HEARTBEAT_INTERVALE_MS)
-}
-
-function startRTTHeartbeat() {
-    if (!rttHeartbeatId) {
-        rttHeartbeatId = setInterval(function () {
-
-            // Send a heartbeat request
-            var request = {
-                operation: "HEARTBEAT",
-                service: "rtt",
-                data: null
-            };
-
-            if (debugEnabled) {
-                console.log("WS SEND: " + JSON.stringify(request));
-            }
-
-            socket.send(JSON.stringify(request));
-        }, 1000 * DEFAULT_RTT_HEARTBEAT);
-    }
 }
 
 function stopHeartbeat(context) {
@@ -265,198 +226,6 @@ function callCallback(context, data, callback) {
     }
 }
 
-function getBrowserName() {
-    // Opera 8.0+
-    var isOpera = (!!window.opr && !!opr.addons) || !!window.opera || (typeof navigator !== 'undefined' && navigator.userAgent.indexOf(' OPR/') >= 0);
-
-    // Firefox 1.0+
-    var isFirefox = typeof InstallTrigger !== 'undefined';
-
-    // Safari 3.0+ "[object HTMLElementConstructor]" 
-    var isSafari = /constructor/i.test(window.HTMLElement) || (function (p) { return p.toString() === "[object SafariRemoteNotification]"; })(!window['safari'] || (typeof safari !== 'undefined' && safari.pushNotification));
-
-    // Internet Explorer 6-11
-    var isIE = (typeof document !== 'undefined' && !!document.documentMode);
-
-    // Edge 20+
-    var isEdge = !isIE && !!window.StyleMedia;
-
-    // Chrome 1+
-    var isChrome = !!window.chrome && !!window.chrome.webstore;
-
-    // Blink engine detection
-    var isBlink = (isChrome || isOpera) && !!window.CSS;
-
-    if (isOpera) return "opera";
-    if (isFirefox) return "firefox";
-    if (isSafari) return "safari";
-    if (isIE) return "ie";
-    if (isEdge) return "edge";
-    if (isChrome) return "chrome";
-    if (isBlink) return "blink";
-
-    return null;
-}
-
-function connect(context, host, port, auth, ssl){
-    rttConnectionStatus = rttConnectionState.CONNECTING;
-    rttAuth = auth
-    
-    // build url with auth as arguments
-    var uri = (ssl ? "wss://" : "ws://") + host + ":" + port;
-    if (rttAuth) {
-        uri += "?";
-        var count = 0;
-        for (var key in rttAuth) {
-            if (count > 0) {
-                uri += "&";
-            }
-            uri += key + "=" + rttAuth[key];
-            ++count;
-        }
-    }
-
-    rttConnectionStatus = rttConnectionState.CONNECTED
-
-    socket = new WebSocket(uri);
-    socket.addEventListener('error', onSocketError);
-    socket.addEventListener('close', onSocketClose);
-    socket.addEventListener('open', onSocketOpen);
-    socket.addEventListener('message', onSocketMessage);
-}
-
-function onSocketOpen(e) {
-    if (exports.isRTTEnabled()) { // This should always be true, but just in case user called disabled and we end up receiving the event anyway
-        
-        // Send a connect request
-        var request = {
-            operation: "CONNECT",
-            service: "rtt",
-            data: {
-                appId: appId,
-                profileId: "s",
-                sessionId: sessionId,
-                system: {
-                    protocol: "ws",
-                    platform: "WEB"
-                }
-            }
-        };
-        
-        var browserName = getBrowserName();
-        if (browserName) {
-            request.data.system.browser = browserName;
-        }
-        
-        request.data.auth = rttAuth;
-
-        if (debugEnabled) {
-            console.log("WS SEND: " + JSON.stringify(request));
-        }
-        socket.send(JSON.stringify(request));
-    }
-}
-
-function rttOnRecv(recv){
-    if (debugEnabled) {
-        console.log("WS RECV: " + JSON.stringify(recv));
-    }
-
-    if (rttCallback) {
-        rttCallback(recv);
-    }
-}
-
-function onSocketMessage(e) {
-    if (exports.isRTTEnabled()) { // This should always be true, but just in case user called disabled and we end up receiving the even anyway
-        var processResult = function (result) {
-            if (result.service == "rtt") {
-                if (debugEnabled) {
-                    console.log("WS RECV: " + JSON.stringify(result));
-                }
-                if (result.operation == "CONNECT") {
-                    DEFAULT_RTT_HEARTBEAT = result.data.heartbeatSeconds; //make default heartbeat match the heartbeat the server gives us
-                    startRTTHeartbeat();
-                
-                    rttConnectCallback.success(result);
-                }
-                else if (result.operation == "DISCONNECT") {
-                    disconnectedWithReason = true;
-                    disconnectMessage =
-                    {
-                        severity: "ERROR",
-                        reason: result.data.reason,
-                        reasonCode: result.data.reasonCode,
-                        data: null
-                    };
-                }
-            }
-            else {
-                rttOnRecv(result);
-            }
-        };
-
-        if (typeof e.data === "string") {
-            processResult(e.data);
-        } else if (typeof FileReader !== 'undefined') {
-            // Web Browser
-            var reader = new FileReader();
-            reader.onload = function () {
-                var parsed = {};
-                try {
-                    parsed = JSON.parse(reader.result);
-                }
-                catch (e) {
-                    console.log("WS RECV: " + reader.result);
-                    parsed = JSON.parse(reader.result); // Trigger the error again and let it fail
-                }
-                processResult(parsed);
-            }
-            reader.readAsText(e.data);
-        } else {
-            // Node.js
-            var parsed = {};
-            try {
-                parsed = JSON.parse(e.data);
-            }
-            catch (e) {
-                console.log("WS RECV: " + e.data);
-                parsed = JSON.parse(e.data); // Trigger the error again and let it fail
-            }
-            processResult(parsed);
-        }
-    }
-}
-
-function onSocketClose(e) {
-    if (exports.isRTTEnabled()) { // Don't spam errors if we get multiple ones
-        rttConnectCallback.failure("close");
-    }
-
-    exports.disableRTT();
-    if (disconnectedWithReason == true) {
-        console.log("RTT:Disconnect" + JSON.stringify(disconnectMessage));
-    }
-}
-
-function onSocketError(e) {
-    if (exports.isRTTEnabled()) { // Don't spam errors if we get multiple ones
-        rttConnectCallback.failure("error");
-    }
-
-    exports.disableRTT();
-}
-
-function requestClientConnection(context, callback) {
-    let json = {
-        service: "rttRegistration",
-        operation: "REQUEST_SYSTEM_CONNECTION",
-        data: {}
-    }
-
-    request(context, json, callback)
-}
-
 /*
  * Create a new S2S context
  * @param appId Application ID
@@ -548,88 +317,44 @@ exports.request = (context, json, callback) => {
     queueRequest(context, json, callback);
 }
 
+/**
+ * Attempts to establish an RTT connection to the brainCloud servers.
+ * @param {*} context object containing session data (appId, serverName, etc.)
+ * @param {*} success function to be invoked when an RTT connection has been established
+ * @param {*} failure function to be invoked if an RTT connection is not established
+ * @returns if RTT is already enabled
+ */
 exports.enableRTT = (context, success, failure) => {
-    disconnectedWithReason = false
-
-    if(exports.isRTTEnabled() || rttConnectionStatus == rttConnectionState.CONNECTING){
-        console.log("RTT is already connected/connecting")
-
-        return;
-    }
-    else{
-        rttConnectCallback = {
-            success: success,
-            failure: failure
-        }
-
-        rttConnectionStatus = rttConnectionState.CONNECTING
-
-        requestClientConnection(context, (context, result) => {
-            if (context.logEnabled) {
-                console.log(JSON.stringify(result))
-            }
-            if (result.status == 200) {
-                for (var i = 0; i < result.data.endpoints.length; ++i) {
-                    var endpoint = result.data.endpoints[i];
-                    if (endpoint.protocol === "ws") {
-                        appId = context.appId
-                        sessionId = context.sessionId
-                        connect(context, endpoint.host, endpoint.port, result.data.auth, endpoint.ssl);
-                        return;
-                    }
-                }
-    
-                // We didn't find websocket endpoint
-                result.status = 0;
-                result.status_message = "WebSocket endpoint missing";
-                rttConnectionStatus = rttConnectionState.DISCONNECTED;
-                rttConnectCallback.failure(result);
-            }
-            else {
-                rttConnectionStatus = rttConnectionState.DISCONNECTED;
-                rttConnectCallback.failure(result);
-            }
-        })
-    }
+    RTT.enableRTT(context, success, failure)
 }
 
+/**
+ * Disables the RTT connection.
+ * @returns if RTT is not enabled
+ */
 exports.disableRTT = () => {
-    if(!(exports.isRTTEnabled()) || rttConnectionStatus == rttConnectionState.DISCONNECTING){
-        console.log("RTT is not enabled")
-        
-        return;
-    }
-    else{
-        rttConnectionStatus = rttConnectionState.DISCONNECTING
-
-        if(rttHeartbeatId){
-            clearInterval(rttHeartbeatId)
-
-            rttHeartbeatId = null
-        }
-
-        if (socket) {
-            socket.removeEventListener('error', onSocketError);
-            socket.removeEventListener('close', onSocketClose);
-            socket.removeEventListener('open', onSocketOpen);
-            socket.removeEventListener('message', onSocketMessage);
-
-            socket.close()
-            socket = null
-        }
-
-        rttConnectionStatus = rttConnectionState.DISCONNECTED
-    }
+    RTT.disableRTT()
 }
 
+/**
+ * Returns whether or not RTT is enabled.
+ * @returns True if RTT is enabled
+ */
+exports.rttIsEnabled = () => {
+    return RTT.rttIsEnabled()
+}
+
+/**
+ * Registers a callback for all RTT services
+ * @param {*} callback function to be invoked when receiving RTT updates
+ */
 exports.registerRTTRawCallback = (callback) => {
-    rttCallback = callback
+    RTT.registerRTTRawCallback(callback)
 }
 
+/**
+ * Deregisters the RTT callback.
+ */
 exports.deregisterRTTRawCallback = () => {
-    rttCallback = null
-}
-
-exports.isRTTEnabled = () => {
-    return rttConnectionStatus == rttConnectionState.CONNECTED
+    RTT.deregisterRTTRawCallback()
 }
